Create speed.py
To calculate the speed we are defining a polygon known as the region of impact(ROI). And then the speed is calculated by the normal speed formula distance divided by the time. But here, the distance and time from the given input video are calculated as follows. First, we calculate the speed by marking the y1 and y2 points of the particular vehicles in different frames and multiplying with the correct calibration factor the time is calculated as the number of frames the vehicle changes to shift from point y1 to point y2 divided by the number of changing per second.